Added css and Javascript
Added colors, fonts, and background color.
Added logic:
A user enters values.
The values are validated.
If the user selected that the baook is read,
then the book will appear on the read column,
if selected not read on the not read column.
The D button will delete the book.
The R button will delete the book and create
a new book with the read status changed to true.